# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 首页
      tags:
      - 首页
  /user/createUser:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 用户姓名
        in: formData
        name: name
        required: true
        type: string
      - description: 输入密码
        in: formData
        name: password
        required: true
        type: string
      - description: 再次输入密码
        in: formData
        name: repassword
        required: true
        type: string
      - description: 电话号码
        in: formData
        name: phone
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: json
      summary: 创建用户
      tags:
      - 用户接口
  /user/deleteUser:
    delete:
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: code", "message"}
          schema:
            type: json
      summary: 删除用户
      tags:
      - 用户接口
  /user/getUserInfo:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 用户姓名
        in: formData
        name: name
        required: true
        type: string
      - description: 输入密码
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: json
      summary: 用户信息
      tags:
      - 用户接口
  /user/getUserList:
    get:
      responses:
        "200":
          description: code", "message"}
          schema:
            type: json
      summary: 用户列表
      tags:
      - 用户接口
  /user/updateUser:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 用户ID
        in: formData
        name: id
        required: true
        type: integer
      - description: 用户姓名
        in: formData
        name: name
        required: true
        type: string
      - description: 新密码
        in: formData
        name: newPassword
        required: true
        type: string
      - description: 电话号码
        in: formData
        name: phone
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: json
      summary: 修改用户信息
      tags:
      - 用户接口
swagger: "2.0"